JWT Athentication

+++ backend/controllers/goalsControllers.js
@@ -1,8 +1,9 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 const getGoal = asyncHandler(async (req, res) => {
-  const goals = await Goal.find({});
+  const goals = await Goal.find({ user: req.user.id });
   res.status(200).json(goals);
 });
 
@@ -10,7 +11,7 @@ const createGoal = asyncHandler(async (req, res) => {
   if (!req.body.text) {
     throw new Error("Please introduced valid text");
   }
-  const goal = await Goal.create({ text: req.body.text });
+  const goal = await Goal.create({ user: req.user.id, text: req.body.text });
   res.status(200).json(goal);
 });
 
@@ -20,6 +21,19 @@ const updateGoal = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error("Goal not found");
   }
+
+  const user = await User.findById(req.user.id);
+
+  if (!user) {
+    res.status(400);
+    throw new Error("User not found");
+  }
+
+  if (goal.user.toString() !== user.id) {
+    res.status(400);
+    throw new Error("User not authorized");
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(
     req.params.id,
     {
@@ -36,6 +50,18 @@ const deleteGoal = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error("Goal not found");
   }
+
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(400);
+    throw new Error("User not found");
+  }
+
+  if (goal.user.toString() !== user.id) {
+    res.status(400);
+    throw new Error("User not authorized");
+  }
+
   await goal.deleteOne();
   res.status(200).json({ id: req.params.id });
 });
